Collapse variant names with empty variation strings
If a variation string is empty, do not add an empty field to the variant
name. Our use case is to remove the host type variant for the current
host:

  host_linux_x86   -> host_x86
  host_windows_x86 -> host_windows_x86

Without this change, we would get a double underscore:

  host_linux_x86 -> host__x86

This could lead to duplicate variantNames in degenerate cases, where
two consecutive variants can be empty, and contain the same value.

Change-Id: I380400a1f917e0617044f5fa87ada30071c05ec1